# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Requirements Inspector REST API",
    "version" : "v1"
  },
  "paths" : {
    "/api/context" : {
      "get" : {
        "operationId" : "getContext",
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            },
            "description" : "Context information"
          }
        },
        "summary" : "Returns basic context information of Polarion's extension",
        "tags" : [ "Extension Information" ]
      }
    },
    "/api/inspect/workitem" : {
      "post" : {
        "operationId" : "inspectRequirement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InspectWorkItemParams"
              }
            }
          },
          "description" : "Workitem to inspect",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WorkItemResponse"
                }
              }
            },
            "description" : "Return value with smell descriptions/counts",
            "headers" : {
              "polarion_requirements_inspector_service_version" : {
                "description" : "Version of requirements-inspector service",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "polarion_requirements_inspector_version" : {
                "description" : "Version of requirements-inspector",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "python_version" : {
                "description" : "Version of requirements-inspector service python",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Inspect a single workitem",
        "tags" : [ "Requirements Inspection" ]
      }
    },
    "/api/inspect/workitems" : {
      "post" : {
        "operationId" : "inspectRequirements",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InspectWorkItemsParams"
              }
            }
          },
          "description" : "List of workitems to inspect",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WorkItemResponse"
                  }
                }
              }
            },
            "description" : "Return value with smell descriptions/counts",
            "headers" : {
              "polarion_requirements_inspector_service_version" : {
                "description" : "Version of requirements-inspector service",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "polarion_requirements_inspector_version" : {
                "description" : "Version of requirements-inspector",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "python_version" : {
                "description" : "Version of requirements-inspector service python",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Inspect multiple workitems",
        "tags" : [ "Requirements Inspection" ]
      }
    },
    "/api/version" : {
      "get" : {
        "operationId" : "getVersion",
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Version"
                }
              }
            },
            "description" : "Version information"
          }
        },
        "summary" : "Returns version of Polarion's extension",
        "tags" : [ "Extension Information" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "Context" : {
        "type" : "object",
        "description" : "Represents the context for building URLs related to Polarion services",
        "properties" : {
          "baseUrl" : {
            "type" : "string",
            "description" : "Returns the base URL constructed with the extension context",
            "example" : "/polarion/pdf-exporter"
          },
          "extensionContext" : {
            "type" : "string",
            "description" : "The extension context used as a base for URL construction",
            "example" : "pdf-exporter"
          },
          "restUrl" : {
            "type" : "string",
            "description" : "Returns the REST API URL constructed with the extension context",
            "example" : "/polarion/pdf-exporter/rest"
          },
          "swaggerUiUrl" : {
            "type" : "string",
            "description" : "Returns the Swagger UI URL for the REST API"
          }
        }
      },
      "InspectWorkItemParams" : {
        "type" : "object",
        "properties" : {
          "addFields" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "addMissingLanguage" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "ignoreInspectTitle" : {
            "type" : "boolean"
          },
          "projectId" : {
            "type" : "string"
          }
        }
      },
      "InspectWorkItemsParams" : {
        "type" : "object",
        "properties" : {
          "addFields" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "addMissingLanguage" : {
            "type" : "boolean"
          },
          "ids" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "ignoreInspectTitle" : {
            "type" : "boolean"
          },
          "projectId" : {
            "type" : "string"
          }
        }
      },
      "Version" : {
        "type" : "object",
        "description" : "Details about the software version",
        "properties" : {
          "automaticModuleName" : {
            "type" : "string",
            "description" : "The automatic module name"
          },
          "bundleBuildTimestamp" : {
            "type" : "string",
            "description" : "The build timestamp of the bundle"
          },
          "bundleName" : {
            "type" : "string",
            "description" : "The name of the bundle"
          },
          "bundleVendor" : {
            "type" : "string",
            "description" : "The vendor of the bundle"
          },
          "bundleVersion" : {
            "type" : "string",
            "description" : "The version of the bundle"
          },
          "projectURL" : {
            "type" : "string",
            "description" : "The project URL"
          },
          "supportEmail" : {
            "type" : "string",
            "description" : "Support email for the bundle"
          }
        }
      },
      "WorkItemResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "missingProcessword" : {
            "type" : "boolean"
          },
          "smellComparative" : {
            "type" : "integer",
            "format" : "int32"
          },
          "smellComplex" : {
            "type" : "string"
          },
          "smellDescription" : {
            "type" : "string"
          },
          "smellPassive" : {
            "type" : "integer",
            "format" : "int32"
          },
          "smellWeakword" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}